FEATURE: util added to compile flight data and populate new_main.flight
