--- utils/data/process/compile/flights/main_flights_table/get_country_code_from_country.go
@@ -1,249 +1,0 @@
-
-package main
-
-import (
-    "fmt"
-    "os"
-    "strings"
-)
-
-// Lookup table mapping various country names and territories (in lowercase) to their ISO codes.
-
-var countryCodes = map[string]string{
-    "afghanistan": "AF",
-    "albania": "AL",
-    "algeria": "DZ",
-    "andorra": "AD",
-    "angola": "AO",
-    "antigua and barbuda": "AG",
-    "argentina": "AR",
-    "armenia": "AM",
-    "australia": "AU",
-    "austria": "AT",
-    "azerbaijan": "AZ",
-    "bahamas": "BS",
-    "bahrain": "BH",
-    "bangladesh": "BD",
-    "barbados": "BB",
-    "belarus": "BY",
-    "belgium": "BE",
-    "belize": "BZ",
-    "benin": "BJ",
-    "bhutan": "BT",
-    "bolivia": "BO",
-    "bosnia and herzegovina": "BA",
-    "botswana": "BW",
-    "brazil": "BR",
-    "brunei": "BN",
-    "bulgaria": "BG",
-    "burkina faso": "BF",
-    "burundi": "BI",
-    "cabo verde": "CV",
-    "cambodia": "KH",
-    "cameroon": "CM",
-    "canada": "CA",
-    "central african republic": "CF",
-    "chad": "TD",
-    "chile": "CL",
-    "china": "CN",
-    "colombia": "CO",
-    "comoros": "KM",
-    "congo": "CG",
-    "congo (brazzaville)": "CG",
-    "congo (kinshasa)": "CD",
-    "democratic republic of the congo": "CD",
-    "costa rica": "CR",
-    "cote d'ivoire": "CI",
-    "ivory coast": "CI",
-    "croatia": "HR",
-    "cuba": "CU",
-    "cyprus": "CY",
-    "czech republic": "CZ",
-    "czechia": "CZ",
-    "denmark": "DK",
-    "djibouti": "DJ",
-    "dominica": "DM",
-    "dominican republic": "DO",
-    "ecuador": "EC",
-    "egypt": "EG",
-    "el salvador": "SV",
-    "equatorial guinea": "GQ",
-    "eritrea": "ER",
-    "estonia": "EE",
-    "eswatini": "SZ",
-    "swaziland": "SZ",
-    "ethiopia": "ET",
-    "fiji": "FJ",
-    "finland": "FI",
-    "france": "FR",
-    "gabon": "GA",
-    "gambia": "GM",
-    "georgia": "GE",
-    "germany": "DE",
-    "ghana": "GH",
-    "greece": "GR",
-    "grenada": "GD",
-    "guatemala": "GT",
-    "guinea": "GN",
-    "guinea-bissau": "GW",
-    "guyana": "GY",
-    "haiti": "HT",
-    "honduras": "HN",
-    "hungary": "HU",
-    "iceland": "IS",
-    "india": "IN",
-    "indonesia": "ID",
-    "iran": "IR",
-    "iraq": "IQ",
-    "ireland": "IE",
-    "israel": "IL",
-    "italy": "IT",
-    "jamaica": "JM",
-    "japan": "JP",
-    "jordan": "JO",
-    "kazakhstan": "KZ",
-    "kenya": "KE",
-    "kiribati": "KI",
-    "north korea": "KP",
-    "korea, democratic people's republic of": "KP",
-    "south korea": "KR",
-    "korea, republic of": "KR",
-    "kuwait": "KW",
-    "kyrgyzstan": "KG",
-    "laos": "LA",
-    "latvia": "LV",
-    "lebanon": "LB",
-    "lesotho": "LS",
-    "liberia": "LR",
-    "libya": "LY",
-    "liechtenstein": "LI",
-    "lithuania": "LT",
-    "luxembourg": "LU",
-    "madagascar": "MG",
-    "malawi": "MW",
-    "malaysia": "MY",
-    "maldives": "MV",
-    "mali": "ML",
-    "malta": "MT",
-    "marshall islands": "MH",
-    "mauritania": "MR",
-    "mauritius": "MU",
-    "mexico": "MX",
-    "micronesia": "FM",
-    "federated states of micronesia": "FM",
-    "moldova": "MD",
-    "monaco": "MC",
-    "mongolia": "MN",
-    "montenegro": "ME",
-    "morocco": "MA",
-    "mozambique": "MZ",
-    "myanmar": "MM",
-    "burma": "MM",
-    "namibia": "NA",
-    "nauru": "NR",
-    "nepal": "NP",
-    "netherlands": "NL",
-    "holland": "NL",
-    "new zealand": "NZ",
-    "nicaragua": "NI",
-    "niger": "NE",
-    "nigeria": "NG",
-    "norway": "NO",
-    "oman": "OM",
-    "pakistan": "PK",
-    "palau": "PW",
-    "palestine": "PS",
-    "state of palestine": "PS",
-    "panama": "PA",
-    "papua new guinea": "PG",
-    "paraguay": "PY",
-    "peru": "PE",
-    "philippines": "PH",
-    "poland": "PL",
-    "portugal": "PT",
-    "qatar": "QA",
-    "romania": "RO",
-    "russia": "RU",
-    "rwanda": "RW",
-    "saint kitts and nevis": "KN",
-    "saint lucia": "LC",
-    "saint vincent and the grenadines": "VC",
-    "samoa": "WS",
-    "san marino": "SM",
-    "sao tome and principe": "ST",
-    "saudi arabia": "SA",
-    "senegal": "SN",
-    "serbia": "RS",
-    "seychelles": "SC",
-    "sierra leone": "SL",
-    "singapore": "SG",
-    "slovakia": "SK",
-    "slovenia": "SI",
-    "solomon islands": "SB",
-    "somalia": "SO",
-    "south africa": "ZA",
-    "south sudan": "SS",
-    "spain": "ES",
-    "sri lanka": "LK",
-    "sudan": "SD",
-    "suriname": "SR",
-    "sweden": "SE",
-    "switzerland": "CH",
-    "syria": "SY",
-    "taiwan": "TW",
-    "tajikistan": "TJ",
-    "tanzania": "TZ",
-    "thailand": "TH",
-    "timor-leste": "TL",
-    "east timor": "TL",
-    "togo": "TG",
-    "tonga": "TO",
-    "trinidad and tobago": "TT",
-    "tunisia": "TN",
-    "turkey": "TR",
-    "turkmenistan": "TM",
-    "tuvalu": "TV",
-    "uganda": "UG",
-    "ukraine": "UA",
-    "united arab emirates": "AE",
-    "united kingdom": "GB",
-    "great britain": "GB",
-    "england": "GB",
-    "scotland": "GB",
-    "wales": "GB",
-    "northern ireland": "GB",
-    "united states": "US",
-    "usa": "US",
-    "uruguay": "UY",
-    "uzbekistan": "UZ",
-    "vanuatu": "VU",
-    "vatican city": "VA",
-    "venezuela": "VE",
-    "vietnam": "VN",
-    "yemen": "YE",
-    "zambia": "ZM",
-    "zimbabwe": "ZW",
-}
-
-// Function to retrieve the ISO code by country or territory name.
-// The input is converted to lowercase to ensure case-insensitive matching.
-func getISOCode(country string) string {
-    country = strings.ToLower(country) // Convert input to lowercase for case-insensitive lookup
-    if code, ok := countryCodes[country]; ok {
-        return code // ISO codes are stored as uppercase
-    }
-    return "Unknown" // Return "Unknown" if the country is not found.
-}
-
-/*
-func main() {
-    if len(os.Args) < 2 {
-        fmt.Println("Please provide a country name or territory as an argument.")
-        return
-    }
-    countryName := os.Args[1] // Get the country name from the first command line argument
-    isoCode := getISOCode(countryName)
-    fmt.Printf("ISO code for %s: %s\n", countryName, isoCode)
-}
-*/
-
