+++ utils/data/process/compile/main/go.mod
@@ -1,10 +1,18 @@
-module compile-main-db 
+module compile-main-db
 
 go 1.18
 
 require (
 	github.com/Tris20/FairFareFinder v0.0.1
 	github.com/mattn/go-sqlite3 v1.14.22
+)
+
+require (
+	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/rivo/uniseg v0.4.7 // indirect
+	github.com/schollz/progressbar/v3 v3.14.6 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	golang.org/x/term v0.22.0 // indirect
 )
 
 replace FairFareFinder/utils/time-and-date => ../../../../../utils/time-and-date
